requirements.txt
asgiref==3.8.1
Brotli==1.1.0
celery==5.4.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2.3
django-debug-toolbar==4.4.2
django-redis==5.4.0
django-tenants==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
dj-database-url==2.1.0
et_xmlfile==2.0.0
fonttools==4.58.4
gunicorn==22.0.0
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.6
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.0.7
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3.post1
